fix 3069
